Starting to rename and refactor Explorer to Ecosystem. Issuel #112.
