+++ lib/screens/app/ecosystem/proposals/proposal_details.dart
@@ -6,7 +6,7 @@ import 'package:provider/provider.dart';
 import 'package:seeds/models/models.dart';
 import 'package:seeds/providers/services/eos_service.dart';
 import 'package:seeds/providers/services/http_service.dart';
-import 'package:seeds/screens/app/explorer/proposals/proposal_header_details.dart';
+import 'package:seeds/screens/app/ecosystem/proposals/proposal_header_details.dart';
 import 'package:seeds/widgets/seeds_button.dart';
 import 'package:url_launcher/url_launcher.dart';
