+++ lib/screens/app/ecosystem/proposals/proposal_header_details.dart
