+++ lib/screens/app/ecosystem/proposals/proposals.dart
@@ -3,7 +3,7 @@ import 'package:flutter_toolbox/flutter_toolbox.dart';
 import 'package:seeds/models/models.dart';
 import 'package:seeds/providers/services/http_service.dart';
 import 'package:seeds/providers/services/navigation_service.dart';
-import 'package:seeds/screens/app/explorer/proposals/proposal_header_details.dart';
+import 'package:seeds/screens/app/ecosystem/proposals/proposal_header_details.dart';
 
 class Proposals extends StatefulWidget {
   @override
